feat: integration tests: adds integration test runner and builds for test, dev, and release (#2)
The primary goal was to create a simple integration test workflow. In order to do so,
some basic work was done to allow building into `test`, `dev`, and `release` buckets
to allow some division between the flows.

The integration tests are run using the nodejs engine as they are just sending requests
to nginx and thus gain no benefit from being executed by njs.

This means we *could* have used a more standard test runner.  However it was kept the same in the name of
consistency.  The integrations tests use the node `http` modules to communicate with nginx over a unix socket.

This choice was made since it will be more performant, but primarily so we don't have to worry about ports being
available or not.