first workflow that should build and push the docker image (#17)
feat: Run integration and unit tests in CI

# What
This sets up Github actions workflows to model what a CI pipeline might look like.  The CI pipeline:

1. Builds a docker image to act as the test env.  The image is cached on the hash of the `Dockerfile`.  Cached builds come it at ~30 seconds while a build clocks in around ~3 minutes.  The image is pushed to github packages under the branch name.  **The code is not built into the container**
2. Runs the unit and integration tests in the previously built container. These tests are run in parallel.

## Javier, what about building the release?
I realized that some changes will be necessary to make that smooth, but I want to keep the size of this PR down.  So I will be doing that in another PR.

## Design
The fundamental idea is that the docker container should be the environment in which tests are run in CI.  You can choose to run tests in the container locally as well.  However, since the docker container is NOT build artifact we don't bake the code into it.  This allows faster builds if the `Dockerfile` is not changed.

It's also the idea that you should modify the `Dockerfile` to more closely mimic the setup/modules of your target nginx setup.


## Technical Tidbits
### Github Actions Quirks
Github Actions does two strange things:

1. Overrides `PATH` in the container
2. [Overrides](https://github.com/actions/runner/issues/863) `HOME` in the container

Because of these two things, you'll see some steps in the jobs to return these to the expected values.

### NPM quirks
`npm run` and what npm calls `lifecycle scripts` which are bash scripts run incidentally by npm usually either fail or are run as a different user if `npm` is invoked by root.  Prior to version 7 of the npm cli you could get around this by passing the `--unsafe-perms` flag to maintain the root user when executing scripts.

As of version 7, the scripts are executed as the user that is the [owner of the directory from while the script is being executed.](https://docs.npmjs.com/cli/v7/using-npm/scripts#user) . As a result, you'll see me doing some aggressive `chown`ing in the job steps to make things work and avoid permissions errors. 

This has made a lot of [people unhappy](https://github.com/npm/cli/issues/4811) and is widely regarded as a bad move.  There is currently an [effort to change how this works.](https://github.com/npm/rfcs/issues/546)