move dockerfile out in public (#34)
refactor: remove the nginx and njs binaries from being set up but the framework

What
This change removes the assumption that there is an nginx and njs executable in the `_internal` folder.  It also normalizes docker-based development commands.

Removal of local `nginx` and `njs` commands
This is the final piece of some work I did a while back to make the commands generally be able to use any nginx executable.  I had a script that attempted to build all the dependencies with the idea that any local nginx installation could be unaffected.

Ultimately I decided that it would be a lot of work to maintain the script and that it would likely be buggy for many users.  Since the framework already does a lot, I decided that attempting to manage njs and nginx binaries would just add further complication as well as putting in "magic" that was fragile and ultimately hurts the goal of helping people get familiar with nginx and njs operation

Thus it is up to the user to make sure that both `nginx` and `njs` are available in their `PATH` and that they are configured to their liking.

Making things easier without doing it for them
Still, installing and configuring `nginx` and `njs` can be considered a significant blocker to some.  To this end, I've formalized commands that make the docker-based flow pretty simple:
| Command                           | description                                                                   |
|-----------------------------------|-------------------------------------------------------------------------      |
| `npm run docker:serve`            | Builds the project and makes it available at `http://localhost:8080` using the provided `Dockerfile`          |
| `npm run docker:release`          | Builds the project and creates a folder structure in `./_build/release` using the provided `Dockerfile`       |
| `npm run docker:test:unit`        | Runs the unit tests using the provided `Dockerfile`                           |
| `npm run docker:test:integration` | Runs the integration tests using the provided `Dockerfile`                    |
| `npm run docker:up`               | Starts the docker container.  This must be run before any `docker:` command   |
| `npm run docker:down`             | Stops the docker container                                                    |

For folks who still want to run nginx locally, I've provided detailed instructions that will be fleshed out further as I finalize docs